Geb's offense ability changes his melee attack, refactoring
Geb's offense ability now has the missing functionality where his melee attack has a longer cooldown but does more damage. This required some refactoring in the ability system, namely changing stats to use floats instead of ints and adding new functions the BaseAbilityInfo that trigger when the ability equipped and unequipped.

+++ Assets/Scripts/Player/PlayerBasicMelee.cs
@@ -17,7 +17,7 @@ public class PlayerBasicMelee : MonoBehaviour
     public float attackRange = 0.5f;
     // public int attackDamage = 20;
     public float attackCooldown = 1f;
-    
+
     float cooldownTimer = 0f;
 
     void Awake()
@@ -43,7 +43,7 @@ public class PlayerBasicMelee : MonoBehaviour
 
         foreach (Collider2D enemy in hitEnemies)
         {
-            enemy.GetComponent<ObjectHealth>().TakeDamage(transform, playerStats.GetValue("Damage"));
+            enemy.GetComponent<ObjectHealth>().TakeDamage(transform, (int)playerStats.GetValue("Damage"));
         }
     }
