+++ Assets/Scripts/Player/StatModifier.cs
@@ -21,9 +21,9 @@ public class StatModifier
 
     [field: SerializeField]
     /// Amount that this stat modifier will change the target stat by. Can be changed in the Unity Editor.
-    private int _modValue;
+    private float _modValue;
     /// Public version of _modValue that can be used by scripts.
-    public int ModValue { get { return _modValue; } set { _modValue = value; } }
+    public float ModValue { get { return _modValue; } set { _modValue = value; } }
     // public StatModType ModType { get; set; }
 
     /*public StatModifier(float modValue, StatModType modType)
@@ -33,7 +33,7 @@ public class StatModifier
     }*/
 
     /// Constructor (function the automatically runs when a new StatModifier is created in another script). Initializes mod value to the given value.
-    public StatModifier(int modValue)
+    public StatModifier(float modValue)
     {
         ModValue = modValue;
     }
